requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
python-dateutil==2.8.2
pandas>=2.1.0
matplotlib>=3.8.0
numpy>=1.26.0
scipy>=1.11.0
reportlab>=4.0.9
Pillow>=10.1.0
openpyxl>=3.1.5
